requirements.txt
openai>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.15.0
wordcloud>=1.9.0
jieba>=0.42.1
requests>=2.31.0
python-dotenv>=1.0.0
